remove circular dependecies
