+++ browser-extension/mv3/src/service-worker/services/requestProcessor/handleHeadersOnRedirect.ts
@@ -1,18 +1,16 @@
 import { AJAXRequestDetails, SessionRuleType } from "./types";
-import { requestProcessor } from ".";
+import { findMatchingRule } from "../ruleMatcher";
+import { updateRequestSpecificRules } from "../rulesManager";
+import { Rule } from "common/types";
 
 export const IGNORED_HEADERS_ON_REDIRECT = ["Authorization"];
 
-export const forwardHeadersOnRedirect = async (tabId: number, requestDetails: AJAXRequestDetails) => {
+export const forwardHeadersOnRedirect = async (tabId: number, requestDetails: AJAXRequestDetails, rules: Rule[]) => {
   if (!IGNORED_HEADERS_ON_REDIRECT.some((header) => checkIfHeaderExists(requestDetails.requestHeaders, header))) {
     return;
   }
 
-  const { isApplied, destinationUrl } =
-    requestProcessor.findMatchingRule(
-      [...requestProcessor.cachedRules.redirectRules, ...requestProcessor.cachedRules.replaceRules],
-      requestDetails
-    ) ?? {};
+  const { isApplied, destinationUrl } = findMatchingRule(rules, requestDetails) ?? {};
 
   if (!isApplied) {
     return;
@@ -27,7 +25,7 @@ export const forwardHeadersOnRedirect = async (tabId: number, requestDetails: AJ
 
   const redirectedUrl = destinationUrl;
 
-  return requestProcessor.updateRequestSpecificRules(
+  return updateRequestSpecificRules(
     tabId,
     requestDetails.url,
     {
